LU-12705 build: fix building fail against Power9 little endian
We use "%ll[dux]" for __u64 variable as an input/output modifier,
this may cause building error on some architectures which use "long"
for 64-bit types, for example, Power9 little endian.
Here add necessary typecasting (long long/unsigned long long) to
make the build correct.

Lustre-change: https://review.whamcloud.com/36007
Lustre-commit: 4eddf36ac3607c66c172668b30eb5dcf921e3de4

Test-Parameters: trivial
Reviewed-on: https://review.whamcloud.com/36207
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/utils/lhsmtool_posix.c
@@ -699,7 +699,8 @@ static int ct_copy_data(struct hsm_copyaction_private *hcp, const char *src,
 					CT_TRACE("bandwith control: %lluB/s "
 						 "excess=%llu sleep for "
 						 "%lld.%09lds",
-						 opt.o_bandwidth, excess,
+						 (unsigned long long)opt.o_bandwidth,
+						 (unsigned long long)excess,
 						 (long long)delay.tv_sec,
 						 delay.tv_nsec);
 					last_bw_print = now;
