+++ lustre/utils/obd.c
@@ -58,7 +58,6 @@
 #include <time.h>
 #include <unistd.h>
 #include <limits.h>
-
 #include "obdctl.h"
 #include "lustreapi_internal.h"
 #include <libcfs/util/list.h>
@@ -1653,7 +1652,7 @@ int jt_obd_test_setattr(int argc, char **argv)
         for (i = 1, next_count = verbose; i <= count && shmem_running(); i++) {
 		if (objid >= OBIF_MAX_OID) {
 			fprintf(stderr, "errr: %s: invalid objid '%llu'\n",
-				jt_cmdname(argv[0]), objid);
+				jt_cmdname(argv[0]), (unsigned long long)objid);
 			return -E2BIG;
 		}
 
@@ -1746,7 +1745,7 @@ int jt_obd_destroy(int argc, char **argv)
         for (i = 1, next_count = verbose; i <= count && shmem_running(); i++, id++) {
 		if (id >= OBIF_MAX_OID) {
 			fprintf(stderr, "errr: %s: invalid objid '%llu'\n",
-				jt_cmdname(argv[0]), id);
+				jt_cmdname(argv[0]), (unsigned long long)id);
 			return -E2BIG;
 		}
 
@@ -1888,7 +1887,7 @@ int jt_obd_test_getattr(int argc, char **argv)
         for (i = 1, next_count = verbose; i <= count && shmem_running(); i++) {
 		if (objid >= OBIF_MAX_OID) {
 			fprintf(stderr, "errr: %s: invalid objid '%llu'\n",
-				jt_cmdname(argv[0]), objid);
+				jt_cmdname(argv[0]), (unsigned long long)objid);
 			return -E2BIG;
 		}
 
@@ -2101,7 +2100,7 @@ int jt_obd_test_brw(int argc, char **argv)
 	ostid_set_seq_echo(&data.ioc_obdo1.o_oi);
 	if (objid >= OBIF_MAX_OID) {
 		fprintf(stderr, "errr: %s: invalid objid '%llu'\n",
-			jt_cmdname(argv[0]), objid);
+			jt_cmdname(argv[0]), (unsigned long long)objid);
 		return -E2BIG;
 	}
